requirements.txt
mcp>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
flask>=2.3.0
